+++ Packet++/src/SSLHandshake.cpp
@@ -1484,7 +1484,7 @@ SSLCertificateMessage::SSLCertificateMessage(uint8_t* data, size_t dataLen, SSLH
 			certificateFull = false;
 		}
 
-		LOG_DEBUG("Parsing certificate: pos=%d; len=%d\n", curPos-data, certificateLength);
+		LOG_DEBUG("Parsing certificate: pos=%d; len=%d\n", (int)(curPos-data), (int)certificateLength);
 		SSLx509Certificate* newCert = new SSLx509Certificate(curPos, certificateLength, certificateFull);
 		m_CertificateList.pushBack(newCert);
 
@@ -1506,7 +1506,7 @@ SSLx509Certificate* SSLCertificateMessage::getCertificate(int index)
 {
 	if (index < 0 || index > (int)m_CertificateList.size())
 	{
-		LOG_DEBUG("certificate index out of range: asked for index %d, total size is %d", index, m_CertificateList.size());
+		LOG_DEBUG("certificate index out of range: asked for index %d, total size is %d", index, (int)m_CertificateList.size());
 		return NULL;
 	}
